ask-your-docs/requirements.txt
streamlit
sentence-transformers
scikit-learn
faiss-cpu
pypdf
numpy
transformers
torch
